requirements.txt
# Core (already have these)
streamlit==1.29.0
python-dotenv==1.0.0
openai==1.6.1
sentence-transformers==2.2.2
chromadb==0.4.18
pandas==2.1.4
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
plotly==5.18.0
tqdm==4.66.1
python-dateutil==2.8.2

# New features
googletrans==4.0.0rc1
feedparser==6.0.10
pillow==10.1.0
SpeechRecognition==3.10.0